refactor(security): minor refactoring; reduced complexity in checkHash
Refs #32
