+++ src/Security/Security.php
@@ -16,6 +16,7 @@ namespace Phalcon\Security;
 use Phalcon\Di\InjectionAwareInterface;
 use Phalcon\Di\Traits\InjectionAwareTrait;
 use Phalcon\Http\RequestInterface;
+use Phalcon\Security\Traits\SecurityGettersTrait;
 use Phalcon\Session\ManagerInterface as SessionInterface;
 use Phalcon\Support\Str\Traits\StartsWithTrait;
 
@@ -40,23 +41,12 @@ use function sprintf;
  *     }
  * }
  *```
- *
- * @property int                   $defaultHash
- * @property int                   $numberBytes
- * @property Random                $random
- * @property string|null           $requestToken
- * @property string|null           $token
- * @property string|null           $tokenKey
- * @property string                $tokenKeySessionId
- * @property string                $tokenValueSessionId
- * @property int                   $workFactor
- * @property RequestInterface|null $localRequest
- * @property SessionInterface|null $localSession
  */
 class Security implements InjectionAwareInterface
 {
     use InjectionAwareTrait;
     use StartsWithTrait;
+    use SecurityGettersTrait;
 
     public const CRYPT_DEFAULT    = 0;
     public const CRYPT_BLOWFISH   = 4;
@@ -66,61 +56,6 @@ class Security implements InjectionAwareInterface
     public const CRYPT_MD5        = 3;
     public const CRYPT_SHA256     = 8;
     public const CRYPT_SHA512     = 9;
-
-    /**
-     * @var int
-     */
-    protected int $defaultHash = self::CRYPT_DEFAULT;
-
-    /**
-     * @var int
-     */
-    protected int $numberBytes = 16;
-
-    /**
-     * @var Random
-     */
-    protected Random $random;
-
-    /**
-     * @var string|null
-     */
-    protected ?string $requestToken = null;
-
-    /**
-     * @var string|null
-     */
-    protected ?string $token = null;
-
-    /**
-     * @var string|null
-     */
-    protected ?string $tokenKey = null;
-
-    /**
-     * @var string
-     */
-    protected string $tokenKeySessionId = '$PHALCON/CSRF/KEY$';
-
-    /**
-     * @var string
-     */
-    protected string $tokenValueSessionId = '$PHALCON/CSRF$';
-
-    /**
-     * @var int
-     */
-    protected int $workFactor = 10;
-
-    /**
-     * @var RequestInterface|null
-     */
-    private ?RequestInterface $localRequest = null;
-
-    /**
-     * @var SessionInterface|null
-     */
-    private ?SessionInterface $localSession = null;
 
     /**
      * Security constructor.
@@ -174,11 +109,7 @@ class Security implements InjectionAwareInterface
         $tokenValue = null,
         bool $destroyIfValid = true
     ): bool {
-        /** @var SessionInterface|null $session */
-        $session = $this->getLocalService('session', 'localSession');
-        if (null !== $session && true === empty($tokenKey)) {
-            $tokenKey = $session->get($this->tokenKeySessionId);
-        }
+        $tokenKey = $this->processTokenKey($tokenKey);
 
         /**
          * If tokenKey does not exist in session return false
@@ -187,23 +118,10 @@ class Security implements InjectionAwareInterface
             return false;
         }
 
-        $userToken = $tokenValue;
-        if (null === $tokenValue) {
-            /** @var RequestInterface|null $request */
-            $request = $this->getLocalService('request', 'localRequest');
-
-            /**
-             * We always check if the value is correct in post
-             */
-            if (null !== $request) {
-                /** @var string|null $userToken */
-                $userToken = $request->getPost($tokenKey, 'string');
-            }
-        }
-
         /**
          * The value is the same?
          */
+        $userToken  = $this->processUserToken($tokenKey, $tokenValue);
         $knownToken = $this->getRequestToken();
         if (null === $knownToken || null === $userToken) {
             return false;
@@ -264,146 +182,6 @@ class Security implements InjectionAwareInterface
         $this->requestToken = null;
 
         return $this;
-    }
-
-    /**
-     * Returns the default hash
-     *
-     * @return int
-     */
-    public function getDefaultHash(): int
-    {
-        return $this->defaultHash;
-    }
-
-    /**
-     * Returns a secure random number generator instance
-     *
-     * @return Random
-     */
-    public function getRandom(): Random
-    {
-        return $this->random;
-    }
-
-    /**
-     * Returns a number of bytes to be generated by the openssl pseudo random
-     * generator
-     *
-     * @return int
-     */
-    public function getRandomBytes(): int
-    {
-        return $this->numberBytes;
-    }
-
-    /**
-     * Returns the value of the CSRF token for the current request.
-     *
-     * @return string|null
-     */
-    public function getRequestToken(): ?string
-    {
-        if (true === empty($this->requestToken)) {
-            return $this->getSessionToken();
-        }
-
-        return $this->requestToken;
-    }
-
-    /**
-     * Returns the value of the CSRF token in session
-     *
-     * @return string|null
-     */
-    public function getSessionToken(): ?string
-    {
-        /** @var SessionInterface|null $session */
-        $session = $this->getLocalService('session', 'localSession');
-        if (null !== $session) {
-            return $session->get($this->tokenValueSessionId);
-        }
-
-        return null;
-    }
-
-    /**
-     * Generate a >22-length pseudo random string to be used as salt for
-     * passwords
-     *
-     * @param int $numberBytes
-     *
-     * @return string
-     * @throws Exception
-     */
-    public function getSaltBytes(int $numberBytes = 0): string
-    {
-        while (true) {
-            $safeBytes = $this->random->base64Safe($numberBytes);
-            if ($safeBytes && mb_strlen($safeBytes) >= $numberBytes) {
-                break;
-            }
-        }
-
-        return $safeBytes;
-    }
-
-    /**
-     * Generates a pseudo random token value to be used as input's value in a
-     * CSRF check
-     *
-     * @return string
-     * @throws Exception
-     */
-    public function getToken(): string
-    {
-        if (null === $this->token) {
-            $this->requestToken = $this->getSessionToken();
-            $this->token        = $this->random->base64Safe($this->numberBytes);
-
-            /** @var SessionInterface|null $session */
-            $session = $this->getLocalService('session', 'localSession');
-            if (null !== $session) {
-                $session->set(
-                    $this->tokenValueSessionId,
-                    $this->token
-                );
-            }
-        }
-
-        return $this->token;
-    }
-
-    /**
-     * Generates a pseudo random token key to be used as input's name in a CSRF
-     * check
-     *
-     * @return string|null
-     * @throws Exception
-     */
-    public function getTokenKey(): ?string
-    {
-        if (null === $this->tokenKey) {
-            /** @var SessionInterface|null $session */
-            $session = $this->getLocalService('session', 'localSession');
-            if (null !== $session) {
-                $this->tokenKey = $this->random->base64Safe($this->numberBytes);
-                $session->set(
-                    $this->tokenKeySessionId,
-                    $this->tokenKey
-                );
-            }
-        }
-
-        return $this->tokenKey;
-    }
-
-    /**
-     * @return int
-     */
-    public function getWorkFactor(): int
-    {
-        return $this->workFactor;
     }
 
     /**
@@ -546,7 +324,7 @@ class Security implements InjectionAwareInterface
      *
      * @return RequestInterface|SessionInterface|null
      */
-    private function getLocalService(string $name, string $property)
+    protected function getLocalService(string $name, string $property)
     {
         if (
             null === $this->$property &&
@@ -557,5 +335,48 @@ class Security implements InjectionAwareInterface
         }
 
         return $this->$property;
+    }
+
+    /**
+     * @param string|null $tokenKey
+     *
+     * @return string|null
+     */
+    private function processTokenKey(string $tokenKey = null): ?string
+    {
+        /** @var SessionInterface|null $session */
+        $session = $this->getLocalService('session', 'localSession');
+        if (null !== $session && true === empty($tokenKey)) {
+            $tokenKey = $session->get($this->tokenKeySessionId);
+        }
+
+        return $tokenKey;
+    }
+
+    /**
+     * @param string      $tokenKey
+     * @param string|null $tokenValue
+     *
+     * @return string|null
+     */
+    private function processUserToken(
+        string $tokenKey,
+        string $tokenValue = null
+    ): ?string {
+        $userToken = $tokenValue;
+        if (null === $tokenValue) {
+            /** @var RequestInterface|null $request */
+            $request = $this->getLocalService('request', 'localRequest');
+
+            /**
+             * We always check if the value is correct in post
+             */
+            if (null !== $request) {
+                /** @var string|null $userToken */
+                $userToken = $request->getPost($tokenKey, 'string');
+            }
+        }
+
+        return $userToken;
     }
 }
